Change SortedInt32KeyNode.height from Int32 to Byte.
This decreases the size of the SortedInt32KeyNode on both 32-bit and 64-bit by ~10%.

+++ src/System.Collections.Immutable/src/System/Collections/Immutable/SortedInt32KeyNode.cs
@@ -53,7 +53,7 @@ namespace System.Collections.Immutable
         /// <summary>
         /// The depth of the tree beneath this node.
         /// </summary>
-        private int height;
+        private byte height; // AVL tree height <= ~1.44 * log2(numNodes + 2)
 
         /// <summary>
         /// The left tree.
@@ -94,7 +94,7 @@ namespace System.Collections.Immutable
             this.right = right;
             this.frozen = frozen;
 
-            this.height = 1 + Math.Max(left.height, right.height);
+            this.height = (byte)(1 + Math.Max(left.height, right.height));
         }
 
         /// <summary>
@@ -586,7 +586,7 @@ namespace System.Collections.Immutable
                     this.right = right;
                 }
 
-                this.height = 1 + Math.Max(this.left.height, this.right.height);
+                this.height = (byte)(1 + Math.Max(this.left.height, this.right.height));
                 return this;
             }
         }
